LayoutTests/imported/w3c/web-platform-tests/css/css-grid/grid-model/grid-box-sizing-001-expected.txt


FAIL .grid 1 assert_equals: 
<div class="grid wholeHeight" style="min-width: 200px; right: 0; box-sizing: border-box;" data-expected-width="200" data-expected-height="100">
    <div class="item" data-expected-width="154" data-expected-height="56"></div>
  </div>
height expected 56 but got 0
FAIL .grid 2 assert_equals: 
<div class="grid wholeHeight" style="width: 200px; right: 0; box-sizing: border-box;" data-expected-width="200" data-expected-height="100">
    <div class="item" data-expected-width="154" data-expected-height="56"></div>
  </div>
height expected 56 but got 0
FAIL .grid 3 assert_equals: 
<div class="grid wholeHeight" style="min-width: 200px; right: 0;" data-expected-width="246" data-expected-height="100">
    <div class="item" data-expected-width="200" data-expected-height="56"></div>
  </div>
height expected 56 but got 0
FAIL .grid 4 assert_equals: 
<div class="grid wholeHeight" style="width: 200px; right: 0;" data-expected-width="246" data-expected-height="100">
    <div class="item" data-expected-width="200" data-expected-height="56"></div>
  </div>
height expected 56 but got 0
FAIL .grid 5 assert_equals: 
<div class="grid wholeHeight" style="right: 0; box-sizing: border-box;" data-expected-width="200" data-expected-height="100">
    <div class="item" data-expected-width="154" data-expected-height="56"></div>
  </div>
height expected 56 but got 0
FAIL .grid 6 assert_equals: 
<div class="grid wholeHeight" style="right: 0; box-sizing: border-box;" data-expected-width="200" data-expected-height="100">
    <div class="item" data-expected-width="154" data-expected-height="56"></div>
  </div>
height expected 56 but got 0
FAIL .grid 7 assert_equals: 
<div class="grid wholeHeight" style="right: 0;" data-expected-width="200" data-expected-height="100">
    <div class="item" data-expected-width="154" data-expected-height="56"></div>
  </div>
height expected 56 but got 0
FAIL .grid 8 assert_equals: 
<div class="grid wholeHeight" style="right: 0;" data-expected-width="200" data-expected-height="100">
    <div class="item" data-expected-width="154" data-expected-height="56"></div>
  </div>
height expected 56 but got 0
FAIL .grid 9 assert_equals: 
<div class="grid wholeHeight" style="min-width: 200px; box-sizing: border-box;" data-expected-width="200" data-expected-height="100">
    <div class="item" data-expected-width="154" data-expected-height="56"></div>
  </div>
height expected 56 but got 0
FAIL .grid 10 assert_equals: 
<div class="grid wholeHeight" style="width: 200px; box-sizing: border-box;" data-expected-width="200" data-expected-height="100">
    <div class="item" data-expected-width="154" data-expected-height="56"></div>
  </div>
height expected 56 but got 0
FAIL .grid 11 assert_equals: 
<div class="grid wholeHeight" style="min-width: 200px;" data-expected-width="246" data-expected-height="100">
    <div class="item" data-expected-width="200" data-expected-height="56"></div>
  </div>
height expected 56 but got 0
FAIL .grid 12 assert_equals: 
<div class="grid wholeHeight" style="width: 200px;" data-expected-width="246" data-expected-height="100">
    <div class="item" data-expected-width="200" data-expected-height="56"></div>
  </div>
height expected 56 but got 0
FAIL .grid 13 assert_equals: 
<div class="grid wholeWidth" style="min-height: 100px; bottom: 0; box-sizing: border-box;" data-expected-width="200" data-expected-height="100">
    <div class="item" data-expected-width="154" data-expected-height="56"></div>
  </div>
height expected 56 but got 0
PASS .grid 14 
FAIL .grid 15 assert_equals: 
<div class="grid wholeWidth" style="min-height: 100px; bottom: 0;" data-expected-width="200" data-expected-height="144">
    <div class="item" data-expected-width="154" data-expected-height="100"></div>
  </div>
height expected 100 but got 0
PASS .grid 16 
FAIL .grid 17 assert_equals: 
<div class="grid wholeWidth" style="bottom: 0; box-sizing: border-box;" data-expected-width="200" data-expected-height="100">
    <div class="item" data-expected-width="154" data-expected-height="56"></div>
  </div>
height expected 56 but got 0
FAIL .grid 18 assert_equals: 
<div class="grid wholeWidth" style="bottom: 0; box-sizing: border-box;" data-expected-width="200" data-expected-height="100">
    <div class="item" data-expected-width="154" data-expected-height="56"></div>
  </div>
height expected 56 but got 0
FAIL .grid 19 assert_equals: 
<div class="grid wholeWidth" style="bottom: 0;" data-expected-width="200" data-expected-height="100">
    <div class="item" data-expected-width="154" data-expected-height="56"></div>
  </div>
height expected 56 but got 0
FAIL .grid 20 assert_equals: 
<div class="grid wholeWidth" style="bottom: 0;" data-expected-width="200" data-expected-height="100">
    <div class="item" data-expected-width="154" data-expected-height="56"></div>
  </div>
height expected 56 but got 0
FAIL .grid 21 assert_equals: 
<div class="grid wholeWidth" style="min-height: 100px; box-sizing: border-box;" data-expected-width="200" data-expected-height="100">
    <div class="item" data-expected-width="154" data-expected-height="56"></div>
  </div>
height expected 56 but got 0
PASS .grid 22 
FAIL .grid 23 assert_equals: 
<div class="grid wholeWidth" style="min-height: 100px;" data-expected-width="200" data-expected-height="144">
    <div class="item" data-expected-width="154" data-expected-height="100"></div>
  </div>
height expected 100 but got 0
PASS .grid 24 
